tl: encode/decode arrays of primitive types via array.array; tl: decode int and long with struct.unpack tl: read int types directly (IntX.read instead of SerializationUtils.read(stream, IntX))

+++ tests/conftest.py
@@ -32,10 +32,10 @@ T = TypeVar("T")
 
 async def _custom_auth_create(_) -> bytes:
     from piltover.db.models import AuthKey
-    from piltover.tl import Int
+    from piltover.tl import Long
 
     key = urandom(256)
-    key_id = Int.read_bytes(hashlib.sha1(key).digest()[-8:])
+    key_id = Long.read_bytes(hashlib.sha1(key).digest()[-8:])
     await AuthKey.create(id=str(key_id), auth_key=key)
     return key
